hotfix/ac_cycle_node_if_not_alive
AC off then AC on if a node's IPMI stack is not alive.

*Background*
Previous implementation is pending on PDU feature:
If PDU gets the node in On status, but node is off by externel
operation, PDU can't immediately power on node. It has to off
first to match node's current status then operate node to power
on.